Ninja Forms 3 compatibility
Using Ninja Forms 3 in a popup will currently trigger the undefined
event close listener on line 66, since it does not use a form element.

I’ve added a check for a Ninja Form outside of the check for form
elements, which closes the box when the form submission gets a
response.

I’ve also added a delay following the style of the generic form
tracking, otherwise the box closes to quickly for the success message
to be shown. Perhaps it would be nice to add a configuration option to
set this delay instead of relying on a filter.